Functions

MultiscaleGraphSignalTransforms.Bilinear_rendering — Method

Bilinear_rendering(X, Img_Mat)

NN_RENDERING generates a rendering signal at each point of X from the image Img_Mat by bilinear interpolation method.

MultiscaleGraphSignalTransforms.GraphSig_Plot — Method

GraphSig_Plot(G::GraphSig; symmetric::Bool = false,
  markersize::Float64 = 2.,
  markercolor::Symbol = :balance,
  markershape::Symbol = :circle,
  markerstrokewidth::Float64 = 1.0,
  markerstrokealpha::Float64 = 1.0,
  markervaluevaries::Bool = true,
  linewidth::Float64 = 1.,
  linecolor::Symbol = :blue,
  linestyle::Symbol = :solid,
  clim::Tuple{Float64,Float64} = (0., 0.),
  notitle::Bool = false, nocolorbar::Bool = false, nolegend::Bool = true,
  stemplot::Bool = false, sortnodes::Bool = false)

Display a plot of the data in a GraphSig object

Input Argument

G::GraphSig: an input GraphSig object

symmetric symmetrize the colorbar

markersize the size of the nodes

markercolor markercolor scheme

markershape shape of marker

markerstrokewidth width of marker stroke

markerstrokealpha capacity of marker stroke

markervaluevaries if the marker color depends on the signal value

linewidth the width of the lines in gplot

linecolor the color of the lines (1D) / graph edges (2D & 3D)

linestyle: the style of line

notitle display a title

nocolorbar display a colorbar

nolegend display a legend

stemplot use a stem plot

clim specify the dynamic display range

sortnodes plot the signal values from smallest to largest in magnitude

MultiscaleGraphSignalTransforms.HGLET_dictionary — Method

HGLET_dictionary(GP::GraphPart, G::GraphSig; method::Symbol = :L)

assemble the whole HGLET dictionary

Input Arguments

GP: a GraphPart object

G: a GraphSig object

method: :L or :Lsym

Output Argument

dictionary: the HGLET dictionary

MultiscaleGraphSignalTransforms.K_functional — Method

K_functional(𝐩::Vector{Float64}, 𝐪::Vector{Float64}, 𝚽::Matrix{Float64},
             ∇𝚽::Matrix{Float64}, 𝛌::Vector{Float64}; length::Any = 1,
             T::Any = :Inf, dt::Float64 = 0.5/maximum(𝛌), tol::Float64 = 1e-5)

computes the K_functional between two vector meassures 𝐩 and 𝐪 on a graph.

Input Argument

𝐩::Vector{Float64}: the source vector measure.

𝐪::Vector{Float64}: the destination vector measure.

𝚽::Matrix{Float64}: matrix of the unweighted graph Laplacian eigenvectors.

∇𝚽::Matrix{Float64}: gradient of unweighted graph Laplacian eigenvectors.

𝛌::Vector{Float64}: vector of eigenvalues.

length::Any: vector of edge lengths (default: 1 represents unweighted graphs)

T::Any: the stopping time T in K_functional (default: :Inf)

tol::Float64: tolerance for convergence (default: 1e-5)

Output Argument

K::Float64: TSD distance d_{TSD}(p, q; T).

E::Float64: an estimated upper bound on the absolute error. In general, E <= tol * norm(K).

MultiscaleGraphSignalTransforms.LPHGLET_Analysis_All — Method

function LPHGLET_Analysis_All(G::GraphSig, GP::GraphPart; ϵ::Float64 = 0.3)

For a GraphSig object 'G', generate the 2 matrices of Lapped-HGLET expansion coefficients corresponding to the eigenvectors of L and Lsym

Input Arguments

G: a GraphSig object

GP: a GraphPart object

ϵ: relative action bandwidth (default: 0.3)

Output Argument

dmatrixlH: the matrix of expansion coefficients for L

dmatrixlHsym: the matrix of expansion coefficients for Lsym

GP: a GraphPart object

MultiscaleGraphSignalTransforms.LPHGLET_Synthesis — Method

function LPHGLET_Synthesis(dvec::Vector{Float64}, GP::GraphPart, BS::BasisSpec, G::GraphSig; method::Symbol = :L, ϵ::Float64 = 0.3)

Perform Lapped-HGLET Synthesis transform

Input Arguments

dvec: the expansion coefficients corresponding to the chosen basis

GP: a GraphPart object

BS: a BasisSpec object

G: a GraphSig object

method: :L or :Lsym, indicating which eigenvectors are used

ϵ: relative action bandwidth (default: 0.3)

Output Argument

f: the reconstructed signal

GS: the reconstructed GraphSig object

MultiscaleGraphSignalTransforms.LPHGLET_dictionary — Method

LPHGLET_dictionary(GP::GraphPart, G::GraphSig; method::Symbol = :L, ϵ::Float64 = 0.3)

assemble the whole LP-HGLET dictionary

Input Arguments

GP: a GraphPart object

G: a GraphSig object

method: :L or :Lsym

ϵ: relative action bandwidth (default: 0.3)

Output Argument

dictionary: the LP-HGLET dictionary

MultiscaleGraphSignalTransforms.NGWP_jkl — Method

function NGWP_jkl(GP_star::GraphPart, drow::Int, dcol::Int)

Generate the (j,k,l) indices for the NGWP basis vector corresponding to the coefficient dmatrix(drow,dcol)

Input Arguments

GP_star::GraphPart: a GraphPart object of the dual grpah

drow::Int: the row of the expansion coefficient

dcol::Int: the column of the expansion coefficient

Output Argument

j: the level index of the expansion coefficient

k: the subregion in dual graph's index of the expansion coefficient

l: the tag of the expansion coefficient

MultiscaleGraphSignalTransforms.NN_rendering — Method

NN_rendering(X, Img_Mat)

NN_RENDERING generates a rendering signal at each point of X from the image Img_Mat by nearest neighbor method.

MultiscaleGraphSignalTransforms.ROT_Distance — Method

ROT_Distance(A, B, Q; edge_length = 1, α = 1.0)

computes the ROT distance matrix from A's column vectors to B's column vectors. If A, B are vector inputs, then it also returns the cost value and the optimal transport plan.

Input Argument

A::Any: a vector or matrix whose columns are initial probability measures.

B::Any: a vector or matrix whose columns are terminal probability measures.

Q::SparseMatrixCSC{Int64,Int64}: the unweighted incidence matrix of the graph.

edge_length::Any: the length vector (default: 1 represents unweighted graphs)

α::Float64: default is 1.0. ROT parameter.

retPlan::Bool: an indicator if return the optimal plan (default: false)

Output Argument

dis::Matrix{Float64}: distance matrix, dis[i,j] = d_{ROT}(aᵢ, bⱼ; α).

MultiscaleGraphSignalTransforms.SunFlowerGraph — Method

SunFlowerGraph(; N = 400)

SUNFLOWERGRAPH construct a simple weighted sunflower graph with N vertices. Edge weights are the reciprocal of Euclidean distances.

Input Arguments

N::Int64: default is 400, the number of vertices. Requires N > 26.

Output Argument

G::SimpleWeightedGraph{Int64,Float64}: a simple weighted graph of the sunflower.

L::Matrix{Float64}: the weighted unnormalized graph Laplacian matrix.

X::Matrix{Float64}: a matrix whose i-th row represent the 2D coordinates of the i-th node.

MultiscaleGraphSignalTransforms.alternating_numbers — Method

alternating_numbers(n)

ALTERNATING_NUMBERS e.g., n = 5, returns [1,5,2,4,3]; n = 6, returns [1,6,2,5,3,4]

Input Arguments

n::Int64: number of nodes in x-axis.

Output Argument

arr::Array{Int64}: result array.

MultiscaleGraphSignalTransforms.approx_error_plot — Method

approx_error_plot(DVEC::Vector{Vector{Float64}}; frac::Float64 = 0.50)

draw relative approx. error w.r.t. fraction of coefficients retained (FCR)

Input Arguments

DVEC::Vector{Vector{Float64}}: a list of expansion coefficients.

fraction::Float64: default is 0.5.

MultiscaleGraphSignalTransforms.approx_error_plot2 — Method

approx_error_plot2(DVEC::Vector{Vector{Float64}}; frac::Float64 = 0.50)

draw relative approx. error w.r.t. fraction of coefficients retained (FCR) in dissertation

Input Arguments

DVEC::Vector{Vector{Float64}}: a list of expansion coefficients.

fraction::Float64: default is 0.5.

MultiscaleGraphSignalTransforms.bs_level — Method

BS = bs_level(GP::GraphPart, j::Int, c2f::Bool = true)

Specify the basis corresponding to level j for a given graph partitioning

Input Arguments

GP::GraphPart: an input GraphPart object

j::Int: the level to which the basis corresponds (j = 0 is the global level)

c2f::Bool: a flag for c2f or f2c (default: true, i.e., c2f)

Output Argument

BS::BasisSpec: a BasisSpec object corresponding to the level j basis

MultiscaleGraphSignalTransforms.bs_walsh — Method

BS = bs_walsh(GP::GraphPart)

Specify the walsh basis corresponding to for a given graph partitioning

Input Arguments

GP::GraphPart: an input GraphPart object

Output Argument

BS::BasisSpec: a BasisSpec object corresponding to the walsh basis

MultiscaleGraphSignalTransforms.bsfull — Method

(levlistfull, levlengthsfull, transfull) = bsfull(GP::GraphPart, BS::BasisSpec, trans::Vector{Bool})

Given a BasisSpec object, return the full-length, redundant levlist, levlengths, and trans descriptions.

Input Arguments

GP::GraphPart: an input GraphPart object

BS::BasisSpec: an input BasisSpec object

trans::Vector{Bool}: a specification of the transforms used for the HGLET-GHWT hybrid transform (default: null)

levlengthsp::Bool: a flag to return levlengthsfull (default: false)

transp::Bool: a flag to return transfull (default: false)

Output Arguments

levlistfull::Vector{UInt8}: the full-length, redundant levels list description

levlengthsfull::Vector{UInt8}: the full-length, redundant levels lengths description

transfull::Matrix{Bool}: the full-length, redundant trans description

MultiscaleGraphSignalTransforms.cat_plot — Method

cat_plot(X; marker = nothing, ms = 4)

CAT_PLOT generates a scatter plot figure for cat example, which is for quick viewing of a graph signal within a specific range (i.e., xlims, ylims, zlims). CAT_PLOT!(X; ...) adds a plot to current one.

Input Arguments

X::Matrix{Float64}: 3-dim points.

marker::Array{Float64}: default is nothing. Present different colors given different signal value at each node.

ms::Array{Float64}: default is 4. Present different node sizes given different signal value at each node.

MultiscaleGraphSignalTransforms.characteristic — Method

characteristic(list, N)

CHARACTERISTIC gives the characteristic function in n-dim vector space with values of index in list equal to 1.

Input Arguments

list::Array{Int}: list of indices.

N::Int: dimension of the target vector.

Output Argument

v::Array{Float64}: the n-dim characteristic vector with values of index in list equal to 1.

MultiscaleGraphSignalTransforms.compute_SNR — Method

compute_SNR(f, g)

COMPUTE_SNR, g = f + ϵ, SNR = 20 * log10(norm(f)/norm(g-f)).

Input Arguments

f::Array{Float64}: original signal.

g::Array{Float64}: noisy signal.

Output Argument

SNR::Float64: SNR value.

MultiscaleGraphSignalTransforms.const_proj_wavelets — Method

const_proj_wavelets(𝚽,vlist,elist; method = "Modified Gram-Schmidt with Lp Pivoting")

construct projection wavelets, i.e., project δ ∈ vlist onto span({φⱼ| j ∈ elist}).

Input Arguments

𝚽::Matrix{Float64}: graph Laplacian eigenvectors 𝚽

vlist::Array{Int}: the list of considered node indices.

elist::Array{Int}: the list of considered eigenvector indices.

method::Symbol: default is :MGSLp. other options: :IP (Iterative-Projection), GS (Gram Schmidt).

Output Argument

Wav::Matrix{Float64}: a matrix whose columns are projected wavelet vectors.

MultiscaleGraphSignalTransforms.cost_functional — Method

costfun = cost_functional(cfspec::Any)

Determine the cost functional to be used by the best-basis algorithm.

Input Argument

cfspec::Any: the specification for the cost functional

Output Argument

costfun::Function: the cost functional (as a function_handle)

MultiscaleGraphSignalTransforms.dct1d — Method

dct1d(k, N)

DCT1D returns k-th 1D DCT basis vector in Rᴺ.

Input Arguments

k::Int64: ord of DCT basis vector. k = 1,2,...,N.

N::Int64: vector dimension.

Output Argument

φ::Array{Float64}: k-th 1D DCT basis vector in Rᴺ. (k is 1-indexed)

MultiscaleGraphSignalTransforms.dct2d_basis — Method

dct2d_basis(N1, N2)

DCT2D_BASIS returns 2D DCT basis vectors in [0,1] x [0,1] with N1-1 and N2-1 subintervals respectively.

Input Arguments

N1::Int64: number of nodes in x-axis.

N2::Int64: number of nodes in y-axis.

Output Argument

𝚽::Matrix{Float64}: 2D DCT basis vectors.

MultiscaleGraphSignalTransforms.dmatrix2dvec — Method

dvec = dmatrix2dvec(dmatrix::Array{Float64,3}, GP::GraphPart, BS::BasisSpec)

Given a matrix of expansion coefficients, convert it to a vector. This function assumes that the input coefficient array dmatrix is in the coarse-to-fine format. If BS.c2f == false, then this function internally converts dmatrix into the fine-to-coarse format. Hence, if one supplies the f2c dmatrix, the results become wrong, and the subsequent procedure may result in error.

Input Arguments

dmatrix::Array{Float64,3}: matrices of expansion coefficients

GP::GraphPart: an input GraphPart object

BS::BasisSpec: an input BasisSpec object

Outputs Arguments

dvec::Matrix{Float64}: a vector of expansion coefficients

MultiscaleGraphSignalTransforms.dmatrix2dvec — Method

(dvec, BS) = dmatrix2dvec(dmatrix::Array{Float64,3}, GP::GraphPart)

Given a matrix of expansion coefficients, convert it to a vector.

Input Arguments

dmatrix::Array{Float64,3}: matrices of expansion coefficients

GP::GraphPart: an input GraphPart object

BS::BasisSpec: an input BasisSpec object

Outputs Arguments

dvec::Matrix{Float64}: a vector of expansion coefficients

BS::BasisSpec: an output BasisSpec object

MultiscaleGraphSignalTransforms.dmatrix_flatten — Method

dmatrix_flatten(dmatrix::Array{Float64,3}, flatten::Any)

Flatten dmatrix using the method specified by the string "flatten"

Input Arguments

dmatrix::Array{Float64,3}: the matrix of expansion coefficients; after this function is called, it becomes the size of (~, ~, 1).

flatten::Any: the method for flattening dmatrix (see the code in details)

MultiscaleGraphSignalTransforms.dualgraph — Method

dualgraph(dist::Matrix{Float64}; method::Symbol = :inverse, σ::Float64 = 1.0)

build the dual graph's weight matrix based on the given non-trivial eigenvector metric.

Input Arguments

dist::Matrix{Float64}: eigenvector distance matrix

method::Symbol: default is by taking inverse of the distance between eigenvectors. Ways to build the dual graph edge weights. Option: :inverse, :gaussian.

σ::Float64: default is 1.0. Gaussian variance parameter.

Output Argument

G_star::GraphSig: A GraphSig object containing the weight matrix of the dual graph.

MultiscaleGraphSignalTransforms.dvec2dmatrix — Method

dmatrix = dvec2dmatrix(dvec::Matrix{Float64}, GP::GraphPart, BS::BasisSpec)

Given a vector of expansion coefficients, convert it to a matrix.

Input Arguments

dvec::Matrix{Float64}: a vector of expansion coefficients

GP::GraphPart: a GraphPart object

BS::BasisSpec: a BasisSpec object

Output Argument

dmatrix::Array{Float64,3}: a set of matrices of expansion coefficients

MultiscaleGraphSignalTransforms.dvec_Threshold — Method

function dvec_Threshold(dvec::Array{Float64}, SORH::String, keep::Float64, GP::GraphPart, BS::BasisSpec)

Threshold HGLET / GHWT coefficients

Input Arguments

dvec::Array{Float64,2} the vector of expansion coefficients

SORH::String use soft ('s') or hard ('h') thresholding

keep::Float64 a fraction between 0 and 1 which says how many coefficients should be kept

GP::GraphPart a GraphPart object, used to identify scaling coefficients

BS::BasisSpec a BasisSpec object, used to identify scaling coefficients

Output Argument

dvec_new::Vector{Float64} the thresholded expansion coefficients

kept::Float64 the number of coefficients kept

MultiscaleGraphSignalTransforms.eigDAG_Distance — Method

eigDAG_Distance(𝚽, Q, numEigs; edge_weights = 1)

compute DAG distances between pairwise graph Laplacian eigenvectors.

Input Arguments

𝚽::Matrix{Float64}: matrix of graph Laplacian eigenvectors, 𝜙ⱼ₋₁ (j = 1,...,size(𝚽,1)).

Q::Matrix{Float64}: incidence matrix of the graph.

numEigs::Int64: number of eigenvectors considered.

edge_weight::Any: default value is 1, stands for unweighted graph (i.e., all edge weights equal to 1). For weighted graph, edge_weight is the weights vector, which stores the affinity weight of each edge.

Output Argument

dis::Matrix{Float64}: a numEigs x numEigs distance matrix, dis[i,j] = d_DAG(𝜙ᵢ₋₁, 𝜙ⱼ₋₁).

MultiscaleGraphSignalTransforms.eigHAD_Affinity — Method

eigHAD_Affinity(𝚽, 𝛌; indexEigs = 1:size(𝚽,2))

EIGHAD_AFFINITY compute Hadamard (HAD) affinity between pairwise graph Laplacian eigenvectors.

Input Arguments

𝚽::Matrix{Float64}: matrix of graph Laplacian eigenvectors, 𝜙ⱼ₋₁ (j = 1,...,size(𝚽,1)).

𝛌::Array{Float64}: array of eigenvalues. (ascending order)

indexEigs::Int: default is all eigenvectors, indices of eigenvectors considered.

Output Argument

A::Matrix{Float64}: a numEigs x numEigs affinity matrix, A[i,j] = a_HAD(𝜙ᵢ₋₁, 𝜙ⱼ₋₁).

MultiscaleGraphSignalTransforms.eigHAD_Distance — Method

eigHAD_Distance(𝚽, 𝛌; indexEigs = 1:size(𝚽,2))

compute HAD "distance" (not really a distance) between pairwise graph Laplacian eigenvectors, i.e., dHAD(𝜙ᵢ₋₁, 𝜙ⱼ₋₁) = √(1 - aHAD(𝜙ᵢ₋₁, 𝜙ⱼ₋₁)²).

Input Arguments

𝚽::Matrix{Float64}: matrix of graph Laplacian eigenvectors, 𝜙ⱼ₋₁ (j = 1,...,size(𝚽,1)).

𝛌::Array{Float64}: array of eigenvalues. (ascending order)

indexEigs::Int: default is all eigenvectors, indices of eigenvectors considered.

Output Argument

dis::Matrix{Float64}: the HAD distance matrix, dis[i,j] = d_HAD(𝜙ᵢ₋₁, 𝜙ⱼ₋₁).

MultiscaleGraphSignalTransforms.eigROT_Distance — Method

eigROT_Distance(P, Q; edge_length = 1, α = 1.0)

computes the ROT distance matrix of P's column vectors on a graph.

Input Argument

P::Matrix{Float64}: a matrix whose columns are vector measures with the same total mass.

Q::SparseMatrixCSC{Int64,Int64}: the unweighted incidence matrix of the graph.

edge_length::Any: the length vector (default: 1 represents unweighted graphs)

α::Float64: default is 1.0. ROT parameter.

Output Argument

dis::Matrix{Float64}: distance matrix, dis[i,j] = d_{ROT}(pᵢ, pⱼ; α).

MultiscaleGraphSignalTransforms.eigTSD_Distance — Method

eigTSD_Distance(P::Matrix{Float64}, 𝚽::Matrix{Float64}, 𝛌::Vector{Float64},
                Q::SparseMatrixCSC{Int64,Int64}; length::Any = 1,
                T::Any = :Inf, tol::Float64 = 1e-5)

computes the TSD distance matrix of P's column vectors on a graph.

Input Argument

P::Matrix{Float64}: vector measures with the same total mass 0.

𝚽::Matrix{Float64}: matrix of the unweighted graph Laplacian eigenvectors.

𝛌::Vector{Float64}: vector of eigenvalues.

Q::SparseMatrixCSC{Int64,Int64}: the unweighted incidence matrix.

length::Any: vector of edge lengths (default: 1 represents unweighted graphs)

T::Any: the stopping time T in K_functional (default: :Inf)

tol::Float64: tolerance for integral convergence (default: 1e-5)

Output Argument

dis::Matrix{Float64}: distance matrix, d_{TSD}(φᵢ, φⱼ; T).

MultiscaleGraphSignalTransforms.eigsROT_Distance — Method

eigsROT_Distance(P::Matrix{Float64}, W::SparseMatrixCSC{Float64, Int64}, X::Matrix{Float64}; α::Float64 = 1.0)

computes the sROT distance matrix of P's column vectors on a (unweighted) tree.

Input Argument

P::Matrix{Float64}: a matrix whose columns are vector measures with the same total mass.

W::SparseMatrixCSC{Float64, Int64}: the weight matrix of the tree.

X::Matrix{Float64}: the node positions (i-th row represents node i's location)

α::Float64: default is 1.0. ROT parameter.

Output Argument

dist_sROT::Matrix{Float64}: distance matrix, distsROT[i,j] = d{sROT}(pᵢ, pⱼ; α).

Ws::SparseMatrixCSC{Float64, Int64}: the weight matrix of the simplified tree.

Xs::Matrix{Float64}: the node locations of the simplified tree

𝚯::Matrix{Float64}: the shortened pmfs from P.

MultiscaleGraphSignalTransforms.find_subgraph_inds — Method

find_subgraph_inds(Wc::SparseMatrixCSC{Float64, Int64})

find all subgraph indices of a tree. (subgraph includes: branches and junctions)

Input Argument

Wc::SparseMatrixCSC{Float64, Int64}: the weight matrix of the tree chopped by the junctions.

Output Argument

Ind::Matrix{Float64}: a matrix whose columns represent the subgraph node indices in binary form.

MultiscaleGraphSignalTransforms.findminimum — Method

findminimum(v, n)

FINDMINIMUM finds the first n smallest elements' indices.

Input Arguments

v::Array{Float64}: the candidate values for selection.

n::Int: number of smallest elements for consideration.

Output Argument

idx::Array{Int}: n smallest elements' indices.

MultiscaleGraphSignalTransforms.frame_approx — Method

frame_approx(f, U, V; num_kept = length(f))

approximate signal f by the frame U.

Input Arguments

f::Vector{Float64}: input graph signal

U::Matrix{Float64}: a frame operator (matrix or dictionary)

V::Matrix{Float64}: the dual frame operator of U

num_kept::Int64: number of kept coefficients (NCR)

Output Argument

rel_error::Vector{Float64}: the relative errors

f_approx::Vector{Float64}: the approximated signal

MultiscaleGraphSignalTransforms.freq_band_matrix — Method

freq_band_matrix(ls,n)

FREQ_BAND_MATRIX provides characteristic diagonal matrix, which is useful for spectral graph filters design.

Input Arguments

ls::Array{Int}: list of indices.

n::Int: dimension of the target vector.

Output Argument

D::Array{Float64}: the zero/one diagonal matrix.

MultiscaleGraphSignalTransforms.getall_expansioncoeffs — Method

getall_expansioncoeffs(G_Sig::GraphSig, GP_star::GraphPart, VM_NGWP::Array{Float64,3}, PC_NGWP::Array{Float64,3}, 𝚽::Matrix{Float64})

get all expansion coefficients of f via all methods in NGWP.jl and MTSG.jl

Input Arguments

G_Sig::GraphSig: GraphSig of the primal graph

GP_star::GraphPart: GraphPart of the dual graph

VM_NGWP::Array{Float64,3}: varimax NGWP.

PC_NGWP::Array{Float64,3}: pair-clustering NGWP.

𝚽::Matrix{Float64}: the matrix of graph Laplacian eigenvectors.

MultiscaleGraphSignalTransforms.getall_expansioncoeffs2 — Method

getallexpansioncoeffs2(GSig::GraphSig, GPstar::GraphPart, GPstarLsym::GraphPart, VMNGWP::Array{Float64,3}, PCNGWP::Array{Float64,3}, LPNGWP::Array{Float64,3}, VMNGWPLsym::Array{Float64,3}, PCNGWPLsym::Array{Float64,3}, LPNGWPLsym::Array{Float64,3}, 𝚽::Matrix{Float64}, 𝚽sym::Matrix{Float64})

get all expansion coefficients of f in dissertation via all methods in NGWP.jl and MTSG.jl

Input Arguments

G_Sig::GraphSig: GraphSig of the primal graph

GP_star::GraphPart: GraphPart of the dual graph

VM_NGWP::Array{Float64,3}: varimax NGWP.

PC_NGWP::Array{Float64,3}: pair-clustering NGWP.

LP_NGWP::Array{Float64,3}: lapped NGWP.

VM_NGWP_Lsym::Array{Float64,3}: Lsym version of varimax NGWP.

PC_NGWP_Lsym::Array{Float64,3}: Lsym version of pair-clustering NGWP.

LP_NGWP_Lsym::Array{Float64,3}: Lsym version of lapped NGWP.

𝚽::Matrix{Float64}: the matrix of L eigenvectors.

𝚽sym::Matrix{Float64}: the matrix of Lsym eigenvectors.

MultiscaleGraphSignalTransforms.gplot — Method

gplot(A, xyz; plotp = true,
      style = :auto, width = 2, color = :blue,
      shape = :none, mwidth = 2, mcolor = color, malpha = 1.0,
      mscolor = color, mswidth = 1, msalpha = 1.0,
      grid::Bool = false, label = "", subplot = 1)

GPLOT Plot graph, as in graph theory. GPLOT(A,xyz,...) plots the graph specified by the adjacency matrix A and the node coordinates xyz. GPLOT!(A,xyz,...) adds a plot to current one.

Input Arguments

A::SparseMatrixCSC{Float64,Int}: the adjacency matrix of a graph G

xyz::Matrix{Float64}: The coordinates array, xyz, is an n-by-2 or n-by-3 matrix

with the position for node i in the i-th row, xyz[i,:] = [x[i] y[i]] or xyz[i,:] = [x[i] y[i] z[i]].

plotp::Bool: if the plot is made (default) or return the X, Y, Z arrays

style::Symbol: of line; choose from Symbol[:auto,:solid,:dash,:dot,:dashdot] (default: :auto)

width::Number: of line in pixels (default: 2; fraction, e.g., 0.5 is allowed)

color::Symbol: of line; choose from Symbol[:white,:blue,:red,:green,...] (default: :blue)

shape::Symbol: choose from Symbol[:none,:auto,:circle,:rect,:star5,:diamond, :hexagon,:cross,:xcross,:utriangle, :dtriangle,:pentagon,:heptagon,:octagon, :star4,:star6,:star7,:star8,:vline,:hline] (default: :none)

mwidth::Number: marker size (or radius) in pixels (default: 2)

mcolor::Symbol: of marker (default: the same as line color)

malpha::Float64: opacity of marker interior; choose from [0,1] (default: 1.0)

mswidth::Number: marker stroke size (width) in pixels (default: 1)

mscolor::Symbol: of marker stroke (default: the same as line color)

msalpha::Float64: opacity of marker stroke; choose from [0,1] (default: 1.0)

grid::Bool: a flag to show grid lines (default: false)

label::String: a string for legend (default: "")

subplot::Int: subplot index (default: 1)

Output Arguments

X::Matrix{Float64}: Nan-punctuated X coordinate vector

Y::Matrix{Float64}: Nan-punctuated Y coordinate vector

Z::Matrix{Float64}: Nan-punctuated Z coordinate vector

(X,Y)= GPLOT(A,xyz,plotp=false,...) or (X,Y,Z) = GPLOT(A,xyz,plotp=false,...) return the NaN-punctuated vectors X and Y or X, Y and Z without actually generating a plot. These vectors can be used to generate the plot at a later time with PLOT or PLOT3D if desired.

A backward-compatible elaboration of Mathworks's gplot that uses 3D data (if available) when the plot is rotated. Robert Piche, Tampere Univ. of Tech., 2005

Translated into julia by Naoki Saito, Dec. 21, 2016. Note that we should still consider the keywords organized as function gplot(A,xyz,kw...)

Nicholas Hausch edits: Must install Plots package For 3D rotation, use plotlyjs() backend

Revised by Naoki Saito, Feb. 17, 2017 Revised by Naoki Saito, Oct. 13, 2017 Revised by Haotian Li, Jul. 25, 2018 Revised by Haotian Li and Naoki Saito for Julia v0.7/1.0, Sep. 21, 2018 Instead of plot(...), we decided to use Plots.plot(...) in order to avoid the function name ambiguity; this is a safer approach. Revised by Naoki Saito, Oct. 22, 2018

MultiscaleGraphSignalTransforms.heat_sol — Method

heat_sol(f0,𝚽,Σ,t)

HEAT_SOL gives the solution of heat partial differential equation with initial condition u(⋅, 0) = f0

Input Arguments

f0::Array{Float64}: initial condition vector.

𝚽::Matrix{Float64}: graph Laplacian eigenvectors, served as graph Fourier transform matrix

Σ::Array{Int}: diagonal matrix of eigenvalues.

t::Float: time elapse.

Output Argument

u::Array{Float64}: the solution vector at time t

MultiscaleGraphSignalTransforms.ind_class — Method

T = ind_class(N::Int)

Given `N`, determine what class `ind` (and `rs`) should be

Input Argument

N::Int: the length of the graph signal

Output Argument

T <: Unsigned: the unsigned integer class that ind should be

Copyright 2015 The Regents of the University of California

Implemented by Jeff Irion (Adviser: Dr. Naoki Saito) | Translated and revised by Naoki Saito, Feb. 8, 2017

MultiscaleGraphSignalTransforms.levlist2levlengths! — Method

levlist2levlengths!(GP::GraphPart, BS::BasisSpec)

Compute the levlengths info for a BasisSpec object

Input Arguments

GP::GraphPart: a GraphPart object

BS::BasisSpec: BasisSpec object, without levlengths; after this function, the levlengths field is filled.

MultiscaleGraphSignalTransforms.lp_ngwp — Method

lp_ngwp(𝚽::Matrix{Float64}, W_dual::SparseMatrixCSC{Float64, Int64},
        GP_dual::GraphPart; ϵ::Float64 = 0.3)

construct the lapped NGWP and GP.tag in place.

Input Arguments

𝚽::Matrix{Float64}: graph Laplacian eigenvectors

W_dual::SparseMatrixCSC{Float64, Int64}: weight matrix of the dual graph

GP_dual::GraphPart: GraphPart object of the dual graph

Output Argument

wavelet_packet::Array{Float64,3}: the lapped NGWP dictionary. The first index is for selecting wavelets at a fixed level; the second index is for selecting the level j; the third index is for selecting elements in the wavelet vector.

MultiscaleGraphSignalTransforms.lp_ngwp_analysis — Method

lp_ngwp_analysis(G::GraphSig, 𝚽::Matrix{Float64}, W_dual::SparseMatrixCSC{Float64, Int64},
        GP_dual::GraphPart; ϵ::Float64 = 0.3)

perform the LP-NGWP transform of the graph signal(s) G.f.

Input Arguments

G::GraphSig: a GraphSig object

𝚽::Matrix{Float64}: graph Laplacian eigenvectors

W_dual::SparseMatrixCSC{Float64, Int64}: weight matrix of the dual graph

GP_dual::GraphPart: GraphPart object of the dual graph

Output Argument

wavelet_packet::Array{Float64,3}: the lapped NGWP dictionary. The first index is for selecting wavelets at a fixed level; the second index is for selecting the level j; the third index is for selecting elements in the wavelet vector.

MultiscaleGraphSignalTransforms.mgslp — Method

mgslp(A::Matrix{Float64}; tol::Float64 = 1e-12, p::Float64 = 1.0)

Modified Gram-Schmidt Process Orthogonalization with ℓᵖ pivoting algorithm (MGSLp)

Input Arguments

A::Matrix{Float64}: whose column vectors are to be orthogonalized.

Output Argument

A::Matrix{Float64}: orthogonalization matrix of A's column vectors based on ℓᵖ pivoting.

MultiscaleGraphSignalTransforms.nat_spec_filter — Method

nat_spec_filter(l, D; σ = 0.25 * maximum(D), method = :regular, thres = 0.2)

assemble the natural spectral graph filter centered at the l-th eigenvector via the distance matrix D.

Input Arguments

l::Int64: index of the centered eigenvector

D::Matrix{Float64}: non-trivial distance matrix of the eigenvectors

σ::Float64: Gaussian window width parameter (default: 0.25 * maximum(D))

method::Symbol: :regular or :reduced (default: :regular)

thres::Float64: cutoff threshold ∈ (0, 1).

Output Argument

𝛍::Vector{Float64}: the natural spectral graph filter

MultiscaleGraphSignalTransforms.natural_eigdist — Method

natural_eigdist(𝚽, 𝛌, Q; α = 1.0, T = :Inf,
                input_format = :zero_measures, distance = :DAG,
                edge_weight = 1, edge_length = 1)

compute natural distances between graph Laplacian eigenvectors.

Input Arguments

𝚽::Matrix{Float64}: matrix of (weighted) graph Laplacian eigenvectors.

𝛌::Vector{Float64}: vector of eigenvalues.

Q::Matrix{Float64}: unweighted incidence matrix of the graph.

α::Float64: ROT parameter. (default: 1.0)

T::Any: TSD parameter, i.e., the stopping time T in K_functional (default: :Inf)

input_format::Symbol: options: :zero_measures, :pmf1 and :pmf2 (default: :zero_measures)

distance::Symbol: options: :ROT, :HAD, :DAG and :TSD (default: :DAG)

edg_length::Any: vector of edge lengths (default: 1 represents unweighted graphs)

edge_weight::Any: the weights vector, which stores the affinity weight of each edge (default: 1 represents unweighted graphs).

Output Argument

dis::Matrix{Float64}: the distance matrix, dis[i,j] = d(𝜙ᵢ₋₁, 𝜙ⱼ₋₁).

MultiscaleGraphSignalTransforms.ngwf_all_vectors — Method

ngwf_all_vectors(D, 𝚽; σ = 0.2 * maximum(D))

assemble the whole NGWF dictionary.

Input Arguments

D::Matrix{Float64}: non-trivial distance matrix of the eigenvectors

𝚽::Matrix{Float64}: graph Laplacian eigenvectors

σ::Float64: Gaussian window width parameter (default: 0.25 * maximum(D))

Output Argument

𝓤::Matrix{Float64}: the NGWF dictionary

MultiscaleGraphSignalTransforms.ngwp_analysis — Method

ngwp_analysis(G::GraphSig, wavelet_packet::Array{Float64,3})

For a GraphSig object G, generate the matrix of NGWP expansion coefficients.

Input Arguments

G::GraphSig: an input GraphSig object

wavelet_packet::Array{Float64,3}: the varimax wavelets packet.

Output Argument

dmatrix::Array{Float64,3}: the expansion coefficients matrix.

MultiscaleGraphSignalTransforms.ngwp_bestbasis — Method

(dvec_ngwp, BS_ngwp) = ngwp_bestbasis(dmatrix::Array{Float64,3}, GP_star::GraphPart;
                                      cfspec::Any = 1.0, flatten::Any = 1.0,
                                      j_start::Int = 1, j_end::Int = size(dmatrix, 2),
                                      useParent::Bool = true)

Select the best basis from the matrix of NGWP expansion coefficients.

Input Arguments

dmatrix::Array{Float64,3}: the matrix of expansion coefficients

GP_star::GraphPart: an input GraphPart object of the dual graph

cfspec::Any: the specification of cost functional to be used (default = 1.0, i.e., 1-norm)

flatten::Any: the method for flattening vector-valued data to scalar-valued data (default = 1.0, i.e, 1-norm)

useParent::Bool: the flag to indicate if we update the selected best basis subspace to the parent when parent and child have the same cost (default = false)

Output Arguments

dvec_ngwp::Matrix{Float64}: the vector of expansion coefficients corresponding to the NGWP best basis

BS_ngwp::BasisSpec: a BasisSpec object which specifies the NGWP best basis

MultiscaleGraphSignalTransforms.nonorth2relerror — Method

function nonorth2relerror(nonorth::array{Float64,1},B::array{Float64,2})

Given a vector 'nonorth' of non-orthonormal expansion coefficients and the matrix 'B' such that B*nonorth is the original signal, return a vector of relative approximation errors when retaining the 1,2,...,N largest coefficients in magnitude.

Input argument

nonorth: a vector of non-orthonormal expansion coefficients

B: the matrix whose such that B*nonorth is the original signal

Output argument

relerror: a vector of relative approximation errors

MultiscaleGraphSignalTransforms.orth2relerror — Method

function orth2relerror(orth)

Given a vector 'orth' of orthonormal expansion coefficients, return a vector of relative approximation errors when retaining the 1,2,...,N largest coefficients in magnitude.

Input argument

orth: a vector of orthonormal expansion coefficients

Output argument

relerror: a vector of relative approximation errors

MultiscaleGraphSignalTransforms.pairclustering — Method

pairclustering(𝚽::Matrix{Float64}, GP_star::GraphPart)

construct the GraphPart object of the primal graph via pair-clustering algorithm.

Input Arguments

𝚽::Matrix{Float64}: graph Laplacian eigenvectors 𝚽

GP_star::GraphPart: GraphPart object of the dual graph

Output Argument

GP::GraphPart: GraphPart object of the primal graph via pair-clustering

MultiscaleGraphSignalTransforms.partition_fiedler — Method

(pm,v) = partition_fiedler(W; method, v)

Partition the vertices of a graph according to the Fiedler vector

Input Arguments

W::SparseMatrixCSC{Float64,Int}: the edge weight matrix

method::Symbol: the parition method to be used (:L or :Lrw or else ...)

v::Vector{Float64}: the Fiedler vector supplied if v is a null vector

Output Arguments

pm::Vector{Int}: a vector of 1's and -1's

v::Vector{Float64}: the Fiedler vector used for graph partitioning

Copyright 2015 The Regents of the University of California

Implemented by Jeff Irion (Adviser: Dr. Naoki Saito) Translated and revised by Naoki Saito, Feb. 9, 2017

MultiscaleGraphSignalTransforms.partition_fiedler_pm — Method

(pm, v) = partition_fiedler_pm(v::Vector{Float64})

Partition an input Graph based on the Fiedler vector `v`

Input Argument

v::Vector{Float64}: the input Fiedler vector

Output Arguments

pm::Vector{Int}: the partition info

v::Vector{Float64}: the output Fiedler vector if requested

Copyright 2015 The Regents of the University of California

Implemented by Jeff Irion (Adviser: Dr. Naoki Saito) Translated and revised by Naoki Saito, Feb. 10, 2017

MultiscaleGraphSignalTransforms.partition_fiedler_troubleshooting — Method

pm = partition_fiedler_troubleshooting(pm,v,W,val)

Troubleshoot potential issues with the partitioning

Input Arguments

pm::Vector{Int}: an input partition info (+1 or -1) of each node

v::Vector{Float64}: the Fiedler vector of the graph

W::SparseMatrixCSC{Float64,Int}: an input edge weight matrix

val::Float64: the algebraic connectivity (i.e., $\lambda_2$)

Ouput Arguments

pm::Vector{Int}: a final partition info vector

Copyright 2015 The Regents of the University of California

Implemented by Jeff Irion (Adviser: Dr. Naoki Saito) Translated and revised by Naoki Saito, Feb. 10, 2017

MultiscaleGraphSignalTransforms.partition_fiedler_troubleshooting — Method

pm = partition_fiedler_troubleshooting(pm)

Troubleshoot potential issues with the partitioning

Input Arguments

pm::Vector{Int}: an input partition info (+1 or -1) of each node

Ouput Arguments

pm::Vector{Int}: a final partition info vector

Copyright 2015 The Regents of the University of California

Implemented by Jeff Irion (Adviser: Dr. Naoki Saito) Translated and revised by Naoki Saito, Feb. 10, 2017

MultiscaleGraphSignalTransforms.partition_pc — Method

partition_pc(𝚽::Matrix{Float64}, indrs::Vector{Int}, indrs_dual1::Vector{Int}, indrs_dual2::Vector{Int})

based on the partition on dual graph, partition the primal graph region (indexed by indrs) into two pieces via the pair-clustering algorithm.

Input Arguments

𝚽::Matrix{Float64}: graph Laplacian eigenvectors 𝚽

indrs::Vector{Int}: indices of the primal graph region to be partitioned

indrs_dual1::Vector{Int}: index1 of the dual graph region's partition result

indrs_dual2::Vector{Int}: index2 of the dual graph region's partition result

Output Argument

indp1::Vector{Int}: index1 of indrs for the partition result

indp2::Vector{Int}: index2 of indrs for the partition result

MultiscaleGraphSignalTransforms.pc_ngwp — Method

pc_ngwp(𝚽::Matrix{Float64}, GP_star::GraphPart, GP::GraphPart)

construct pair-clustering NGWP and GP.tag in place.

Input Arguments

𝚽::Matrix{Float64}: graph Laplacian eigenvectors 𝚽

GP_star::GraphPart: GraphPart object of the dual graph

GP::GraphPart: GraphPart object of the primal graph

Output Argument

wavelet_packet::Array{Float64,3}: the pair-clustering NGWP. The first index is for selecting wavelets at a fixed level; the second index is for selecting the level j; the third index is for selecting elements in the wavelet vector.

MultiscaleGraphSignalTransforms.rngwf_all_vectors — Method

rngwf_all_vectors(D, 𝚽; σ = 0.2 * maximum(D), thres = 0.2)

assemble the reduced NGWF (rNGWF) dictionary.

Input Arguments

D::Matrix{Float64}: non-trivial distance matrix of the eigenvectors

𝚽::Matrix{Float64}: graph Laplacian eigenvectors

σ::Float64: Gaussian window width parameter (default: 0.25 * maximum(D))

thres::Float64: cutoff threshold ∈ (0, 1).

Output Argument

𝓤::Matrix{Float64}: the rNGWF dictionary

dic_l2x::Dict: a dictionary to store the filtered locations by QR at the l-th centered eigenvector

MultiscaleGraphSignalTransforms.rngwf_lx — Method

rngwf_lx(dic_l2x)

find the sequential subindices of rNGWF vectors.

Input Arguments

dic_l2x::Dict: a dictionary to store the filtered locations by QR at the l-th centered eigenvector

Output Argument

Γ::Vector{Tuple{Int64,Int64}}: the sequential subindices of rNGWF vectors.

MultiscaleGraphSignalTransforms.rs_to_region — Method

function rs_to_region(rs::Matrix{Any}, tag::Matrix{Any})

From the imformation of rs, tag from GP, compute the tag_r matrix, which have the same size of dmatrix (expansion coefficient matrix). Each element indicates the place of the coefficient in the expansion tree.

Input Arguments

rs::Matrix{Any}: rs from GP, showing information of the partition tree

tag::Matrix{Any}: tag from GP, indicating coefficients tag

Output Arguments

tag_r::Matrix{UInt64}: showing information of the partition tree, same size as dmatrix

MultiscaleGraphSignalTransforms.scatter_gplot — Method

scatter_gplot(X; marker = nothing, ms = 4, plotOrder = :normal, c = :viridis, subplot = 1)

SCATTER_GPLOT generates a scatter plot figure, which is for quick viewing of a graph signal. SCATTER_GPLOT!(X; ...) adds a plot to current one.

Input Arguments

X::Matrix{Float64}: points locations, can be 2-dim or 3-dim.

marker::Array{Float64}: default is nothing. Present different colors given different signal value at each node.

ms::Array{Float64}: default is 4. Present different node sizes given different signal value at each node.

plotOrder::Symbol: default is normal. Optional choices :s2l or :l2s, i.e., plots from the smallest value of marker to the largest value or the other way around.

c::Symbol: default is :viridis. Colors.

subgplot::Int: default is 1. The subplot index.

MultiscaleGraphSignalTransforms.sort_wavelets — Method

sort_wavelets(A; onlyByLoc = false)

sort A's column wavelet vectors based on their focused nodes' indices. flip signs via the cross correlation.

Input Arguments

A::Matrix{Float64}: whose column vectors are wavelets.

Output Argument

A::Matrix{Float64}: a matrix with sorted and sign flipped column.

MultiscaleGraphSignalTransforms.spectral_clustering — Method

spectral_clustering(𝚽, M)

SPECTRAL_CLUSTERING return M graph clusters, i.e., {Vₖ| k = 1,2,...,M}.

Input Argument

𝚽::Matrix{Float64}: the matrix of graph Laplacian eigenvectors.

M::Int: the number of graph clusters.

Output Argument

clusters::Vector{Vector{Int}}: graph cluster indices.

MultiscaleGraphSignalTransforms.spike — Method

spike(i, N)

SPIKE gives the N-dim spike vector with i-th element equals 1.

Input Arguments

i::Int: index for one.

N::Int: dimension of the target spike vector.

Output Argument

v::Array{Float64}: the N-dim spike vector with i-th element equals 1.

MultiscaleGraphSignalTransforms.standardize_eigenvectors! — Method

standardize_eigenvectors!(𝚽::Matrix{Float64})

standardize the signs of the eigenvectors such that 1) the term with the largest magnitude is positive; 2) if the max == -min, then make sure the first non-zero entry of the eigenvector is positive.

Input Arguments

𝚽::Matrix{Float64}: matrix of graph Laplacian eigenvectors

MultiscaleGraphSignalTransforms.tag_class — Method

T = tag_class(jmax::Int)

Given jmax, determine what class tag should be.

Input Argument

jmax::Int: the number of levels in the recursive partitioning (j = 1:jmax)

Output Argument

T <: Unsigned: the unsigned integer class that tag should be

Copyright 2015 The Regents of the University of California

Implemented by Jeff Irion (Adviser: Dr. Naoki Saito) | Translated and revised by Naoki Saito, Feb. 8, 2017

MultiscaleGraphSignalTransforms.transform2D — Method

transform2D(X; s = 1, t = [0,0])

TRANSFORM2D dilate each point of X by scale s and translate by 2D vector t.

MultiscaleGraphSignalTransforms.varimax — Method

varimax(A; gamma = 1.0, minit = 20, maxit = 1000, reltol = 1e-12)

VARIMAX perform varimax (or quartimax, equamax, parsimax) rotation to the column vectors of the input matrix.

Input Arguments

A::Matrix{Float64}: input matrix, whose column vectors are to be rotated. d, m = size(A).

gamma::Float64: default is 1. gamma = 0, 1, m/2, and d(m - 1)/(d + m - 2), corresponding to quartimax, varimax, equamax, and parsimax.

minit::Int: default is 20. Minimum number of iterations, in case of the stopping criteria fails initially.

maxit::Int: default is 1000. Maximum number of iterations.

reltol::Float64: default is 1e-12. Relative tolerance for stopping criteria.

Output Argument

B::Matrix{Float64}: output matrix, whose columns are already been rotated.

Implemented by Haotian Li, Aug. 20, 2019

MultiscaleGraphSignalTransforms.vm_ngwp — Method

vm_ngwp(𝚽::Matrix{Float64}, GP_star::GraphPart)

construct varimax NGWP and GP_star.tag in place.

Input Arguments

𝚽::Matrix{Float64}: graph Laplacian eigenvectors 𝚽

GP_star::GraphPart: GraphPart object of the dual graph

Output Argument

wavelet_packet::Array{Float64,3}: the varimax NGWP. The first index is for selecting wavelets at a fixed level; the second index is for selecting the level j; the third index is for selecting elements in the wavelet vector.